python-ocr-service/requirements.txt
Flask
flask-cors
python-dotenv
paddlepaddle
paddleocr
opencv-python
ultralytics